chore(file): move mock classes to separate package, fixes
